chore: move app into own directory
