+++ .github/workflows/app.yaml
@@ -1,4 +1,4 @@
-name: umbra
+name: app
 
 on: [pull_request, push]
