--- .vscode/extensions.json
@@ -1,9 +1,0 @@
-{
-  // See https://go.microsoft.com/fwlink/?LinkId=827846
-  // for the documentation about the extensions.json format
-  "recommendations": [
-    "dart-code.dart-code",
-    "dart-code.flutter",
-    "felixangelov.bloc"    
-  ]
-}
