+++ app/analysis_options.yaml
@@ -1,0 +1,4 @@
+include: package:very_good_analysis/analysis_options.3.0.0.yaml
+linter:
+  rules:
+    public_member_api_docs: false
