+++ app/test/app/view/app_test.dart
@@ -5,9 +5,9 @@
 // license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
+import 'package:app/app/app.dart';
+import 'package:app/counter/counter.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:umbra/app/app.dart';
-import 'package:umbra/counter/counter.dart';
 
 void main() {
   group('App', () {
