+++ app/test/counter/cubit/counter_cubit_test.dart
@@ -5,10 +5,9 @@
 // license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
+import 'package:app/counter/counter.dart';
 import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter_test/flutter_test.dart';
-
-import 'package:umbra/counter/counter.dart';
 
 void main() {
   group('CounterCubit', () {
